Function version in 27s on password dataset.
Version tested for both local (same JVM) and remote workers.
Password characters have to be passed as String because List<String> of
'one character String' cannot be deserialized by Kryos.
New variables have to be created when sending by batches because the
messages' variables are passed by reference when the actors are on the
same JVM (leading to ConcurrentAccessError and aberent behavior if the
reference is modified before being processed)